requirement.txt
certifi==2018.4.16
click==6.7
Flask==1.0.2
Flask-Cors==3.0.4
imutils==0.3.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mkl-fft==1.0.0
mkl-random==1.0.1
numpy==1.11.3
olefile==0.45.1
pandas==0.21.1
Pillow==4.2.1
psycopg2==2.7.4
python-dateutil==2.7.2
pytz==2018.4
six==1.11.0
SQLAlchemy==1.2.7
Werkzeug==0.14.1
zbar==0.10
